requirements.txt

Scrapy==2.8.0
sublist3r==1.0
requests==2.31.0
builtwith==1.3.4
python-whois==0.8.0
dnspython==2.4.2
